First bug: training tensors need to int32/int64 for embedding indexing

+++ machine_translation/data_pipeline/collator.py
@@ -31,7 +31,7 @@ class DynamicTrimmingCollator:
         tgt_batch_final[:, -1] = self.eos_id
 
         src_tensors = torch.from_numpy(src_batch_trimmed).long()    # Convert to int64
-        tgt_tensors = torch.from_numpy(tgt_batch_final)             # Convert to int64
+        tgt_tensors = torch.from_numpy(tgt_batch_final).long()      # Convert to int64
 
         batch = Batch(src=src_tensors, tgt=tgt_tensors, pad=self.pad_id)
